refactored sessions and session result

+++ app/services/F1data.sql
@@ -214,7 +214,7 @@ CREATE TABLE session_result (
     meeting_key INTEGER NOT NULL REFERENCES meetings(meeting_key),
     driver_number INTEGER NOT NULL,
     position INTEGER,
-    points NUMERIC(4, 1),
+    points INTEGER,
     dnf BOOLEAN DEFAULT FALSE,
     dns BOOLEAN DEFAULT FALSE,
     dsq BOOLEAN DEFAULT FALSE,
@@ -223,6 +223,7 @@ CREATE TABLE session_result (
     number_of_laps INTEGER,
     UNIQUE(session_key, driver_number)
 );
+
 CREATE TABLE starting_grid (
     id SERIAL PRIMARY KEY,
     session_key INTEGER NOT NULL REFERENCES sessions(session_key),
